@@ -23,13 +23,12 @@
         "test:watch": "nwb test-react --server"
     },
     "dependencies": {
-        "@svgr/webpack": "^5.4.0",
         "autosuggest-highlight": "^3.2.0",
         "clsx": "^1.0.4",
         "jwt-decode": "^3.0.0",
-        "memoize-one": "^5.1.1",
+        "memoize-one": "^6.0.0",
         "oidc-client": "^1.10.1",
-        "prop-types": "15.7.2",
+        "prop-types": "^15.7.2",
         "react-csv-downloader": "^2.7.0",
         "react-request-fullscreen": "^1.1.2",
         "react-virtualized": "^9.21.2"
@@ -42,11 +41,10 @@
         "@mui/material": "^5.5.0",
         "@mui/styles": "^5.5.0",
         "notistack": "^2.0.3",
-        "react-router": "^5.1.2",
-        "react-router-dom": "^5.1.2",
-        "react": "^17.0.0",
-        "react-dom": "^17.0.0",
-        "react-intl": "^5.0.0"
+        "react-router-dom": "^6.0.0",
+        "react": "^18.0.0",
+        "react-dom": "^18.0.0",
+        "react-intl": "^6.0.0"
     },
     "devDependencies": {
         "@emotion/react": "^11.8.1",
@@ -55,28 +53,28 @@
         "@mui/lab": "^5.0.0-alpha.72",
         "@mui/material": "^5.5.0",
         "@mui/styles": "^5.5.0",
+        "@babel/helper-builder-react-jsx": "7.15.4",
+        "@svgr/webpack": "^5.4.0",
         "babel-eslint": "^10.1.0",
         "bufferutil": "^4.0.1",
         "eslint": "^7.9.0",
         "eslint-config-prettier": "^8.0.0",
         "eslint-config-react-app": "^6.0.0",
-        "eslint-loader": "^4.0.2",
+        "eslint-webpack-plugin": "^2.0.0",
         "eslint-plugin-flowtype": "^5.2.0",
         "eslint-plugin-import": "^2.22.0",
         "eslint-plugin-jsx-a11y": "^6.3.1",
         "eslint-plugin-prettier": "^4.0.0",
         "eslint-plugin-react": "^7.21.0",
         "eslint-plugin-react-hooks": "^4.1.2",
-        "inferno": "^7.4.2",
         "notistack": "^2.0.3",
         "nwb": "^0.25.x",
         "prettier": "^2.0.0",
-        "puppeteer": "^10.0.0",
-        "react": "^17.0.0",
-        "react-dom": "^17.0.0",
-        "react-intl": "^5.0.0",
-        "react-router": "^5.1.2",
-        "react-router-dom": "^5.1.2",
+        "puppeteer": "^17.0.0",
+        "react": "^18.0.0",
+        "react-dom": "^18.0.0",
+        "react-intl": "^6.0.0",
+        "react-router-dom": "^6.0.0",
         "type-fest": "^2.0.0",
         "utf-8-validate": "^5.0.2",
         "webpack-plugin-serve": "^1.0.1"
