@@ -6,7 +6,7 @@
  */
 
 import React, { useCallback } from 'react';
-import { Redirect, Route, Switch } from 'react-router-dom';
+import { Navigate, Route, Routes } from 'react-router-dom';
 import SignInCallbackHandler from '../SignInCallbackHandler';
 import {
     handleSigninCallback,
@@ -20,12 +20,12 @@ const AuthenticationRouter = ({
     userManager,
     signInCallbackError,
     dispatch,
-    history,
+    navigate,
     location,
 }) => {
     const handleSigninCallbackClosure = useCallback(
-        () => handleSigninCallback(dispatch, history, userManager.instance),
-        [dispatch, history, userManager.instance]
+        () => handleSigninCallback(dispatch, navigate, userManager.instance),
+        [dispatch, navigate, userManager.instance]
     );
     const handleSilentRenewCallbackClosure = useCallback(
         () => handleSilentRenewCallback(userManager.instance),
@@ -41,35 +41,42 @@ const AuthenticationRouter = ({
                     Error : SignIn Callback Error; {signInCallbackError.message}
                 </h1>
             )}
-            <Switch>
-                <Route exact path="/sign-in-callback">
-                    <SignInCallbackHandler
-                        userManager={userManager.instance}
-                        handleSignInCallback={handleSigninCallbackClosure}
-                    />
-                </Route>
-                <Route exact path="/silent-renew-callback">
-                    <SilentRenewCallbackHandler
-                        userManager={userManager.instance}
-                        handleSilentRenewCallback={
-                            handleSilentRenewCallbackClosure
-                        }
-                    />
-                </Route>
-                <Route exact path="/logout-callback">
-                    <Redirect to="/" />
-                </Route>
-                <Route>
-                    {userManager.error === null && (
-                        <Login
-                            disabled={userManager.instance === null}
-                            onLoginClick={() =>
-                                login(location, userManager.instance)
+            <Routes>
+                <Route
+                    path="sign-in-callback"
+                    element={
+                        <SignInCallbackHandler
+                            userManager={userManager.instance}
+                            handleSignInCallback={handleSigninCallbackClosure}
+                        />
+                    }
+                />
+                <Route
+                    path="silent-renew-callback"
+                    element={
+                        <SilentRenewCallbackHandler
+                            userManager={userManager.instance}
+                            handleSilentRenewCallback={
+                                handleSilentRenewCallbackClosure
                             }
                         />
-                    )}
-                </Route>
-            </Switch>
+                    }
+                />
+                <Route path="logout-callback" element={<Navigate to="/" />} />
+                <Route
+                    path="*"
+                    element={
+                        userManager.error === null && (
+                            <Login
+                                disabled={userManager.instance === null}
+                                onLoginClick={() =>
+                                    login(location, userManager.instance)
+                                }
+                            />
+                        )
+                    }
+                />
+            </Routes>
         </React.Fragment>
     );
 };
